Order processing working (I think)

+++ CatalogService/src/main/kotlin/it/polito/wa2/catalogservice/services/NotificationServiceImpl.kt
@@ -49,7 +49,7 @@ class NotificationServiceImpl : NotificationService {
         }
     }
 
-    @Scheduled(fixedRate = 1000 * 60 * 60 * 24)
+    @Scheduled(fixedRate = (1000 * 60 * 60 * 24).toLong())
     fun removeExpiredTokens() {
         CoroutineScope(Dispatchers.IO).launch {
             removeTokensAndUsers()
